Make functions reusable and simplified them
