+++ src/modules/account/submodule/multisig/views/ViewAccountMultisigEditAccount.vue
@@ -125,7 +125,7 @@ import { networkState } from '@/state/networkState';
 import {useI18n} from 'vue-i18n'
 import { Helper } from '@/util/typeHelper';
 import AccountComponent from "@/modules/account/components/AccountComponent.vue";
-import { TransactionUtils } from '@/util/transactionUtils';
+import { TransactionUtils, fetchAccount, findAcc } from '@/util/transactionUtils';
 import { AppState } from '@/state/appState';
 import AccountTabs from "@/modules/account/components/AccountTabs.vue";
 export default {
@@ -167,12 +167,6 @@ export default {
     const defaultAcc = walletState.currentLoggedInWallet?walletState.currentLoggedInWallet.selectDefaultAccount(): null
     const selectedAccAdd = ref(defaultAcc?defaultAcc.address:'');
     const accBalance = ref(Helper.toCurrencyFormat(defaultAcc?defaultAcc.balance:0, AppState.nativeToken.divisibility));
-     const findAcc = (publicKey)=>{
-      if(!walletState.currentLoggedInWallet){
-        return null
-      }
-      return walletState.currentLoggedInWallet.accounts.find(acc=>acc.publicKey==publicKey)
-    }
     // current wallet
     const wallet = walletState.currentLoggedInWallet;
     // get account details initialization
@@ -402,14 +396,6 @@ export default {
         return '0' 
       }
     });
-    const splitBalance = computed(()=>{
-      let split = initiatorDisplayBalance.value.split(".")
-      if (split[1]!=undefined){
-        return {left:split[0],right:split[1]}
-      }else{
-        return {left:split[0], right:null}
-      }
-    })
     const addCoSig = () => {
       coSign.value.push('');
       showAddressError.value.push(false);
@@ -603,10 +589,6 @@ export default {
       
     });
 
-    const fetchAccount = (publicKey) => {
-      return walletState.currentLoggedInWallet.accounts.find(account => account.publicKey === publicKey);
-    };
-
     const getMultiSigCosigner = computed(() => {
       if(networkState.currentNetworkProfileConfig){
         let cosigners = multiSign.getCosignerInWallet(accounts.value.find(account => account.address == selectedAccAdd.value)?accounts.value.find(account => account.address == selectedAccAdd.value).publicKey:'');
@@ -638,7 +620,6 @@ export default {
       contact,
       toggleContact,
       contactName,
-      splitBalance,
       currentNativeTokenName,
       err,
       disableSend,
