+++ src/modules/metadataTxn/views/ViewUpdateAccountMetadata.vue
@@ -112,7 +112,7 @@ import {useI18n} from 'vue-i18n'
 import { multiSign } from "@/util/multiSignatory";
 import { walletState } from "@/state/walletState";
 import { networkState } from "@/state/networkState";
-import { TransactionUtils } from "@/util/transactionUtils";
+import { TransactionUtils, fetchAccount, findAcc } from "@/util/transactionUtils";
 import { AppState } from '@/state/appState';
 import AccountComponent from "@/modules/account/components/AccountComponent.vue";
 import MetadataInput from '@/modules/metadataTxn/components/MetadataInput.vue'
@@ -375,13 +375,6 @@ export default {
         return (showScopedKeyErr.value==true ||inputScopedMetadataKey.value==''||newValue.value==''||!walletPassword.value.match(passwdPattern)||showBalanceErr.value==true)
     })
 
-    const findAcc = (publicKey)=>{
-      if(!walletState.currentLoggedInWallet){
-        return
-      }
-      return walletState.currentLoggedInWallet.accounts.find(acc=>acc.publicKey==publicKey)
-    }
-
     const accounts = computed(()=>{
       if(!walletState.currentLoggedInWallet){
         return [];
@@ -554,16 +547,6 @@ export default {
       }
     });
 
-    const splitBalance = computed(()=>{
-      let accBalance = Helper.toCurrencyFormat(balance.value, AppState.nativeToken.divisibility)
-      let split = accBalance.split(".")
-      if (split[1]!=undefined){
-        return {left:split[0],right:split[1]}
-      }else{
-        return {left:split[0], right:null}
-      }
-    })
-
     const transactionFee = computed(() => {
       return aggregateFee.value.toString()
     })
@@ -571,10 +554,6 @@ export default {
     const totalFeeFormatted = computed(() => {
       return Helper.amountFormatterSimple(totalFee.value, 0);
     });
-
-    const fetchAccount = (publicKey) => {
-      return walletState.currentLoggedInWallet.accounts.find(account => account.publicKey === publicKey);
-    };
 
     const getMultiSigCosigner = computed(() => {
       if(networkState.currentNetworkProfileConfig){
@@ -655,7 +634,6 @@ export default {
       lockFund,
       aggregateFee,
       lockFundTxFee,
-      splitBalance,
       walletName,
       currentNativeTokenName,
       oldValue,
