+++ src/modules/services/submodule/assets/views/ViewServicesAssetsLinkToNamespace.vue
@@ -95,7 +95,7 @@ import Tooltip from 'primevue/tooltip';
 import { ThemeStyleConfig } from '@/models/stores/themeStyleConfig';
 import { multiSign } from '@/util/multiSignatory';
 import { AppState } from '@/state/appState';
-import { TransactionUtils } from '@/util/transactionUtils';
+import { isMultiSig, TransactionUtils, fetchAccount, findAccWithAddress } from '@/util/transactionUtils';
 
 export default {
   name: 'ViewServicesAssetsLinkToNamespace',
@@ -159,21 +159,6 @@ export default {
     const balance = ref('');
     const balanceNumber = ref(0);
 
-    const isMultiSig = (address) => {
-      if(walletState.currentLoggedInWallet){
-        const account = walletState.currentLoggedInWallet.accounts.find((account) => account.address == address) || walletState.currentLoggedInWallet.others.find((account) => account.address == address);
-        if(account){
-          const isMulti = account.getDirectParentMultisig().length>0?true:false
-          return isMulti
-        }else{
-          return false
-        }
-      }else{
-        return false
-      }
-    };
-
-
     const isMultiSigBool = ref(isMultiSig(selectedAccAdd.value));
 
     const showNoBalance = computed(() => {
@@ -229,10 +214,6 @@ export default {
 
     const transactionFee = ref('0.000000');
     const transactionFeeExact = ref(0);
-
-    const fetchAccount = (publicKey) => {
-      return walletState.currentLoggedInWallet.accounts.find(account => account.publicKey === publicKey);
-    };
 
     const getMultiSigCosigner = computed(()=>{
       if(!account.value){
@@ -369,22 +350,6 @@ export default {
       }
     });
 
-    const splitCurrency = (amount) => {
-      let split = amount.toString().split(".")
-      if (split[1]!=undefined){
-        return '<span class="font-semibold text-sm">' + split[0] + '</span>.<span class="font-semibold text-xs">' + split[1] + ' ' + currentNativeTokenName.value + '</span>';
-      }else{
-        return '<span class="font-semibold text-sm">' + split[0] + '</span> <span class="font-semibold text-xs">' + currentNativeTokenName.value + '</span>';
-      }
-    };
-
-    const findAccWithAddress = address =>{
-      if(!walletState.currentLoggedInWallet){
-        return null
-      }
-      return walletState.currentLoggedInWallet.accounts.find(acc=>acc.address==address)
-    }
-
     const checkCosignBalance = computed(() => {
       let cosignBalance = findAccWithAddress(cosignerAddress.value)?findAccWithAddress(cosignerAddress.value).balance:0;
       return Helper.toCurrencyFormat(cosignBalance);
@@ -408,7 +373,6 @@ export default {
       supply,
       disabledPassword,
       currencyName,
-      isMultiSig,
       isMultiSigBool,
       selectAsset,
       selectAction,
@@ -426,7 +390,6 @@ export default {
       cosigner,
       isNotCosigner,
       disabledSelectAction,
-      splitCurrency,
       Helper,
       svgString,
       checkCosignBalance,
