+++ src/modules/services/submodule/assets/views/ViewServicesAssetsModifySupplyChange.vue
@@ -97,7 +97,7 @@ import Tooltip from 'primevue/tooltip';
 import { ThemeStyleConfig } from '@/models/stores/themeStyleConfig';
 import { multiSign } from '@/util/multiSignatory';
 import { AppState } from '@/state/appState';
-import { TransactionUtils } from '@/util/transactionUtils';
+import { isMultiSig, TransactionUtils, fetchAccount, findAccWithAddress } from '@/util/transactionUtils';
 
 export default {
   name: 'ViewServicesAssetsModifySupplyChange',
@@ -187,19 +187,6 @@ export default {
       return Helper.toCurrencyFormat(account.value.balance, AppState.nativeToken.divisibility)
     })
 
-    const isMultiSig = (address) => {
-      if(walletState.currentLoggedInWallet){
-        const account = walletState.currentLoggedInWallet.accounts.find((account) => account.address == address) || walletState.currentLoggedInWallet.others.find((account) => account.address == address);
-        if(!account){
-          return false
-        }
-        const isMulti = account.getDirectParentMultisig().length>0?true:false
-        return isMulti
-      }else{
-        return false
-      }
-    };
-
     const supply = ref('0');
     const plainAddress = Helper.createAddress(props.address).plain()
     let account = computed(()=>{
@@ -263,12 +250,6 @@ export default {
       }
       return asset.value.supply/Math.pow(10,asset.value.divisibility)
     })
-
-    
-
-    const fetchAccount = (publicKey) => {
-      return walletState.currentLoggedInWallet.accounts.find(account => account.publicKey === publicKey);
-    };
     
     const getMultiSigCosigner = computed(()=>{
       if(!account.value){
@@ -435,21 +416,6 @@ export default {
       }
     });
 
-    const splitCurrency = (amount) => {
-      let split = amount.toString().split(".")
-      if (split[1]!=undefined){
-        return '<span class="font-semibold text-sm">' + split[0] + '</span>.<span class="font-semibold text-xs">' + split[1] + ' ' + currentNativeTokenName.value + '</span>';
-      }else{
-        return '<span class="font-semibold text-sm">' + split[0] + '</span> <span class="font-semibold text-xs">' + currentNativeTokenName.value + '</span>';
-      }
-    };
-
-    const findAccWithAddress = address =>{
-      if(!walletState.currentLoggedInWallet){
-        return null
-      }
-      return walletState.currentLoggedInWallet.accounts.find(acc=>acc.address==address)
-    }
 
     const checkCosignBalance = computed(() => {
       let cosignBalance = findAccWithAddress(cosignerAddress.value)?findAccWithAddress(cosignerAddress.value).balance:0;
@@ -493,7 +459,6 @@ export default {
       cosigner,
       isNotCosigner,
       disabledSelectIncreaseDecrease,
-      splitCurrency,
       Helper,
       svgString,
       maxAmount,
