+++ src/modules/services/submodule/namespaces/views/ViewServicesNamespaceCreate.vue
@@ -71,7 +71,7 @@ import { networkState } from "@/state/networkState";
 import { Helper } from '@/util/typeHelper';
 import { NamespaceUtils } from '@/util/namespaceUtils';
 import { ChainUtils } from '@/util/chainUtils';
-import { TransactionUtils } from '@/util/transactionUtils';
+import { TransactionUtils, isMultiSig, fetchAccount, findAccWithAddress } from '@/util/transactionUtils';
 import { UnitConverter } from '@/util/unitConverter';
 import { TimeUnit } from '@/models/const/timeUnit';
 import { multiSign } from '@/util/multiSignatory';
@@ -195,15 +195,6 @@ export default {
       walletPassword.value.match(passwdPattern) && namespaceName.value.match(namespacePattern) && (!showDurationErr.value) && (!showNoBalance.value) && (!isNotCosigner.value) && !showNamespaceNameError.value && selectNamespace.value
     ));
 
-    const isMultiSig = (address) => {
-      if(walletState.currentLoggedInWallet){
-        const account = walletState.currentLoggedInWallet.accounts.find((account) => account.address == address) || walletState.currentLoggedInWallet.others.find((account) => account.address == address);
-        const isMulti = account.getDirectParentMultisig().length>0?true:false
-        return isMulti
-      }else{
-        return false
-      }
-    };
     const defaultAcc = walletState.currentLoggedInWallet?walletState.currentLoggedInWallet.selectDefaultAccount(): null
     const selectedAccName = ref(defaultAcc?defaultAcc.name:'');
     const selectedAccAdd = ref(defaultAcc?defaultAcc.address:'');
@@ -252,10 +243,6 @@ export default {
 
     const transactionFee = ref(0);
     const transactionFeeExact = ref(0);
-
-    const fetchAccount = (publicKey) => {
-      return walletState.currentLoggedInWallet.accounts.find(account => account.publicKey === publicKey);
-    };
 
     const getMultiSigCosigner = computed(() => {
       if(networkState.currentNetworkProfileConfig){
@@ -425,26 +412,10 @@ export default {
       }
     })
 
-    const findAccWithAddress = address =>{
-      if(!walletState.currentLoggedInWallet){
-        return null
-      }
-      return walletState.currentLoggedInWallet.accounts.find(acc=>acc.address==address)
-    }
-
     const checkCosignBalance = computed(() => {
       let cosignBalance = findAccWithAddress(cosignerAddress.value)?findAccWithAddress(cosignerAddress.value).balance:0;
       return Helper.toCurrencyFormat(cosignBalance);
     })
-
-    const splitCurrency = (amount) => {
-      let split = amount.toString().split(".")
-      if (split[1]!=undefined){
-        return '<span class="font-semibold text-sm">' + split[0] + '</span>.<span class="font-semibold text-xs">' + split[1] + '</span>';
-      }else{
-        return '<span class="font-semibold text-sm">' + split[0] + '</span>';
-      }
-    };
 
     const reservedRootNamespace = computed(()=>{
       if(networkState.currentNetworkProfileConfig){
@@ -631,7 +602,6 @@ export default {
       disabledDuration,
       duration,
       showDurationErr,
-      isMultiSig,
       isMultiSigBool,
       rentalFee,
       rentalFeeCurrency,
@@ -651,7 +621,6 @@ export default {
       cosignerBalanceInsufficient,
       cosignerAddress,
       isNotCosigner,
-      splitCurrency,
       walletState,
       currentNativeTokenName,
       nsRef,
